Update chess-player.ts
Wasn't pushing elements with modified background colours to the changed list. Also, was calling the warn function whether a piece had been selected with the mousedown or not. Preemptively created a threats function to be coded in the future which will highlight squares with pieces hanging. so that they can be moved out of danger or covered by other pieces.